chore: wip rename to input
