+++ admin/src/components/Input/CKEditorInput/configs/base.js
