+++ admin/src/components/Input/CKEditorInput/configs/blockBalloon.ts
