+++ admin/src/components/Input/CKEditorInput/configs/index.ts
