+++ admin/src/components/Input/CKEditorInput/configs/toolbar.ts
