+++ admin/src/components/Input/CKEditorInput/plugins/StrapiMediaLib.js
