+++ admin/src/components/Input/CKEditorInput/plugins/StrapiUploadAdapter.js
