+++ admin/src/components/Input/CKEditorInput/plugins/index.js
