+++ admin/src/components/Input/CKEditorInput/styling.js
