+++ admin/src/components/Input/CKEditorInput/theme/additional.ts
