+++ admin/src/components/Input/CKEditorInput/theme/common.ts
