+++ admin/src/components/Input/CKEditorInput/theme/dark.ts
