+++ admin/src/components/Input/CKEditorInput/theme/index.ts
