+++ admin/src/components/Input/CKEditorInput/theme/light.ts
